Moved the Videos model to videos app
